test: migrate Phase 5 oauth non-provider tests + mark monitoring db_semantics
Move the 15 non-provider OAuth tests (remove-method state changes, profile
auth-method display, login-required redirects, register form renders,
invalid-invite validation) to tests/component using an OAuth-enabled
component app; provider-callback tests stay e2e. Mark the 8 selection-run
monitoring health tests db_semantics (real selection_run_records queries).
